Complete the proof that the SF election specficaition satisfies the mutual majority criterion.
Theorem:
Assume there is a nonempty group of candidates that a majority of voters
prefer to every other candidate not in the group.
Then, some member of the group will win the election.

Proof Sketch:
The SF spec is total, thus it suffices to show that every possible
winner of the election is in the preferred group.  We proceed by induction
on the round structure of the election.

Invariant: at every round of the election at least on member of the
group is not yet eliminated.

In the base case, some candidate achieves a majority of votes.  This
candidate must be in the preferred group; for otherwise, two disjoint
sets of votes would be a majority (those for the winning candidate,
and those for the preferred group), giving a contradiction.

In the inductive case, no candidate achives a majority of continuing votes
and we eliminate some candidate.  There must remain
some candidate in the group not eliminated.  Suppose otherwise: then
by our invariant, before we eliminated there was exactly one candidate
in the preferred group which was not yet eliminated.  Such a candidate
must have a majority of continuing votes, contrary to the assumption
that no candidate has a majority.

QED

Note one tricky point: by the assumption of mutual majority, the majority
that prefers the group always has their ballots continue through every
round of the election. This is because they prefer every candidate in the group
to candidates not in the group, and must thus at least properly select every
candidate in the group.  Therefore the mutual majority voters always constitute
a majority of continuing votes.